requirements.txt
wandb
gin-config
numba
einops
tensorboard==2.15.1
timm==0.9.12
torch==2.1.1
torchaudio==2.1.1
torchmetrics==1.2.1
torchvision==0.16.1